slight refactor for future changes
