+++ src/twitch_commands/ClipCommand.java
@@ -1,4 +1,4 @@
-package commands;
+package twitch_commands;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
