+++ src/twitch_commands/LogEditCommand.java
@@ -1,4 +1,4 @@
-package commands;
+package twitch_commands;
 
 import java.io.IOException;
 import java.util.Set;
@@ -12,9 +12,9 @@ import com.github.twitch4j.common.enums.CommandPermission;
 import aspectibot.TwitchCommand;
 import utils.CommandLog;
 
-public class LogAddCommand implements TwitchCommand {
+public class LogEditCommand implements TwitchCommand {
 
-    private final Logger LOG = LoggerFactory.getLogger(LogAddCommand.class);
+    private final Logger LOG = LoggerFactory.getLogger(LogEditCommand.class);
     
 	public String response(ChannelMessageEvent event) {
 		String[] message = event.getMessage().split(" ");
@@ -30,15 +30,16 @@ public class LogAddCommand implements TwitchCommand {
 				}
 			}
 			
-			try {			
-			    CommandLog.add(command_name, event.getMessage());
-			    
+			try {	
+			    CommandLog.edit(command_name, event.getMessage());
+				return "";
 			} catch(IOException e) {
-			    LOG.error("response: Unable to add " + command_name + " to the json file!");
-				return "@Onteia you fucked up again! Madge";
+			    LOG.error("response: Unable to edit " + command_name + "!");
+				return "";
 			}
 		}
-		return "";	//if user isn't mod
+		
+		return "";
 	}
 
 }
