+++ src/twitch_commands/LurkCommand.java
@@ -1,14 +1,14 @@
-package commands;
+package twitch_commands;
 
 import com.github.twitch4j.chat.events.channel.ChannelMessageEvent;
 
 import aspectibot.TwitchCommand;
 
-public class TwitterCommand implements TwitchCommand {
+public class LurkCommand implements TwitchCommand {
 
 	public String response(ChannelMessageEvent event) {
 		
-		return "Aspect's twitter: https://twitter.com/aspecticor";
+		return "Thank you for lurking, " + event.getUser().getName() + "! Homi";
 		
 	}
