+++ src/twitch_commands/TwitterCommand.java
@@ -1,14 +1,14 @@
-package commands;
+package twitch_commands;
 
 import com.github.twitch4j.chat.events.channel.ChannelMessageEvent;
 
 import aspectibot.TwitchCommand;
 
-public class EmotesCommand implements TwitchCommand {
+public class TwitterCommand implements TwitchCommand {
 
 	public String response(ChannelMessageEvent event) {
 		
-		return "If you're not able to see all emotes, get the BTTV, FFZ, and 7TV extensions!";
+		return "Aspect's twitter: https://twitter.com/aspecticor";
 		
 	}
